[feat] Enable Resizing Image (#23)

+++ src/main/image_loader.ts
@@ -81,20 +81,25 @@ export async function generateResizedCache(
 		);
 		if (files.length === 0) return;
 
-		await !fs.rm(resizedCacheDir, { recursive: true, force: true });
+		if (fsSync.existsSync(resizedCacheDir)) {
+			await !fs.rm(resizedCacheDir, { recursive: true, force: true });
+		}
 		await fs.mkdir(resizedCacheDir, { recursive: true });
 
 		for (const input of files) {
 			const inputPath = path.join(cacheDir, input);
 			const { name } = path.parse(input);
-			const outputPath = path.join(resizedCacheDir, `${name}.png`);
+			const outputPath = path.join(
+				resizedCacheDir,
+				`${name}-${maxWidth}x${maxHeight}.png`,
+			);
 
 			await sharp(inputPath)
 				.resize({ width: maxWidth, height: maxHeight, fit: "inside" })
 				.toFile(outputPath);
 		}
 	} catch (error) {
-		console.error(`Error Recreating "${resizedCacheDir}"`, error);
+		console.error("Error on creating resized images:", error);
 		throw error;
 	}
 }
